+++ src/Application/Command/Type/Update.php
@@ -67,6 +67,12 @@ class Update implements Command
     /** @var bool */
     public $canScanParticipant;
 
+    /** @var bool */
+    public $isPackageRequired = false;
+
+    /** @var bool */
+    public $isPaymentRequired = false;
+
     /**
      * @param Type   $type
      * @param string $locale
@@ -89,6 +95,8 @@ class Update implements Command
         $this->canRemoveMeeting = $type->canRemoveMeeting();
         $this->areAllSheetParticipantsAssignedToMeeting = $type->areAllSheetParticipantsAssignedToMeeting();
         $this->canScanParticipant = $type->canScanParticipant();
+        $this->isPackageRequired = $type->isPackageRequired();
+        $this->isPaymentRequired = $type->isPaymentRequired();
 
         foreach ($type->getEvent()->getLocales() as $eventLocale) {
             $this->translations[$eventLocale] = [