+++ src/Application/Query/Invoice/InvoiceQuery.php
@@ -10,16 +10,14 @@
 
 namespace Proximum\Vimeet\Application\Query\Invoice;
 
+use Proximum\Vimeet\Application\Query\Query;
 use Proximum\Vimeet\Domain\Model\Invoice\Invoice;
 
-class InvoiceQuery
+class InvoiceQuery implements Query
 {
     /** @var Invoice */
     public $invoice;
 
-    /**
-     * @param Invoice $invoice
-     */
     public function __construct(Invoice $invoice)
     {
         $this->invoice = $invoice;