+++ src/Infrastructure/Bundle/InfrastructureBundle/DependencyInjection/Configuration.php
@@ -61,6 +61,7 @@ class Configuration implements ConfigurationInterface
                 ->scalarNode('import_spot_path')->end()
                 ->scalarNode('import_authentication_token_path')->end()
                 ->scalarNode('encrypted_files_path')->end()
+                ->scalarNode('print_invoices_path')->end()
                 ->arrayNode('package')
                     ->children()
                         ->arrayNode('default_labels')