+++ src/Ui/Bundle/AdminBundle/Form/Type/Type/TypeUpdateType.php
@@ -120,6 +120,14 @@ class TypeUpdateType extends AbstractType
                 'canScanParticipant', CheckboxType::class, [
                 'required' => false,
             ])
+            ->add(
+                'isPackageRequired', CheckboxType::class, [
+                'required' => false,
+            ])
+            ->add(
+                'isPaymentRequired', CheckboxType::class, [
+                'required' => false,
+            ])
         ;
     }