+++ tests/Domain/User/Agenda/Version/DiffVerbalizerTest.php
@@ -15,11 +15,14 @@ use Prophecy\Argument;
 use Prophecy\Prophecy\ObjectProphecy;
 use Proximum\Vimeet\Application\Adapter\TranslatorInterface;
 use Proximum\Vimeet\Domain\Model\Event;
+use Proximum\Vimeet\Domain\Model\Meeting\Message;
+use Proximum\Vimeet\Domain\Model\Meeting\Request;
 use Proximum\Vimeet\Domain\Model\MeetingSlot;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\Spot;
 use Proximum\Vimeet\Domain\Model\User;
 use Proximum\Vimeet\Domain\Model\User\Agenda\Version;
+use Proximum\Vimeet\Domain\Repository\Meeting\MessageRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\MeetingSlotRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\SheetRepositoryInterface;
 use Proximum\Vimeet\Domain\Repository\SpotRepositoryInterface;
@@ -43,6 +46,9 @@ class DiffVerbalizerTest extends TestCase
     /** @var ObjectProphecy */
     private $spotRepository;
 
+    /** @var ObjectProphecy|MessageRepositoryInterface */
+    private $messageRepository;
+
     public function setUp()
     {
         $this->diffChecker = $this->prophesize(DiffChecker::class);
@@ -50,6 +56,8 @@ class DiffVerbalizerTest extends TestCase
         $this->sheetRepository = $this->prophesize(SheetRepositoryInterface::class);
         $this->meetingSlotRepository = $this->prophesize(MeetingSlotRepositoryInterface::class);
         $this->spotRepository = $this->prophesize(SpotRepositoryInterface::class);
+        $this->messageRepository = $this->prophesize(MessageRepositoryInterface::class);
+        $this->messageRepository->getUpdateOrDeleteReasonMessageFromRequestIds(Argument::any())->willReturn([]);
     }
 
     public function testVerbalizeDiffNoDiff()
@@ -64,7 +72,8 @@ class DiffVerbalizerTest extends TestCase
             $this->translator->reveal(),
             $this->sheetRepository->reveal(),
             $this->meetingSlotRepository->reveal(),
-            $this->spotRepository->reveal()
+            $this->spotRepository->reveal(),
+            $this->messageRepository->reveal()
         );
         $result = $diffVerbalizer->verbalizeDiff($lastVersion->reveal(), $currentVersion, $locale);
 
@@ -81,18 +90,21 @@ class DiffVerbalizerTest extends TestCase
         $user = $this->prophesize(User::class);
         $locale = 'fr';
         $currentVersion = [];
-        $lastVersion->getVersion()->willReturn([
-            1 => [
-                'request' => 1,
-                'fromSheet' => 2,
-                'toSheet' => 3,
-            ],
-            2 => [
-                'request' => 2,
-                'fromSheet' => 2,
-                'toSheet' => 4,
-            ],
-        ]);
+        $lastVersion->getVersion()->willReturn(
+            [
+                1 => [
+                    'request' => 1,
+                    'fromSheet' => 2,
+                    'toSheet' => 3,
+                ],
+                2 => [
+                    'request' => 2,
+                    'fromSheet' => 2,
+                    'toSheet' => 4,
+                ],
+            ]
+        )
+        ;
         $lastVersion->getEvent()->willReturn($event->reveal());
         $lastVersion->getUser()->willReturn($user->reveal());
         $event->getTimeZone()->willReturn('Europe/Paris');
@@ -102,11 +114,13 @@ class DiffVerbalizerTest extends TestCase
         $this->sheetRepository
             ->getSheetsByUserAndEvent($user->reveal(), $event->reveal())
             ->shouldBeCalled()
-            ->willReturn([2 => $userSheet->reveal()]);
+            ->willReturn([2 => $userSheet->reveal()])
+        ;
         $this->sheetRepository
             ->findByIds([3 => 3, 4 => 4])
             ->shouldBeCalled()
-            ->willReturn([3 => $sheetMet1->reveal(), 4 => $sheetMet2->reveal()]);
+            ->willReturn([3 => $sheetMet1->reveal(), 4 => $sheetMet2->reveal()])
+        ;
         $this->translator
             ->trans(
                 DiffVerbalizer::TRANSLATION_MEETING_DELETED,
@@ -115,7 +129,8 @@ class DiffVerbalizerTest extends TestCase
                 $locale
             )
             ->shouldBeCalled()
-            ->willReturn('Meeting deleted with sheet title given 1');
+            ->willReturn('Meeting deleted with sheet title given 1')
+        ;
         $this->translator
             ->trans(
                 DiffVerbalizer::TRANSLATION_MEETING_DELETED,
@@ -124,7 +139,8 @@ class DiffVerbalizerTest extends TestCase
                 $locale
             )
             ->shouldBeCalled()
-            ->willReturn('Meeting deleted with sheet title given 2');
+            ->willReturn('Meeting deleted with sheet title given 2')
+        ;
 
         $this->meetingSlotRepository->findByIds([])->shouldNotBeCalled();
         $this->spotRepository->getSpotsByIds([])->shouldNotBeCalled();
@@ -135,7 +151,8 @@ class DiffVerbalizerTest extends TestCase
             $this->translator->reveal(),
             $this->sheetRepository->reveal(),
             $this->meetingSlotRepository->reveal(),
-            $this->spotRepository->reveal()
+            $this->spotRepository->reveal(),
+            $this->messageRepository->reveal()
         );
         $result = $diffVerbalizer->verbalizeDiff($lastVersion->reveal(), $currentVersion, $locale);
 
@@ -171,15 +188,18 @@ class DiffVerbalizerTest extends TestCase
                 'spot' => 14,
             ],
         ];
-        $lastVersion->getVersion()->willReturn([
-            1 => [
-                'request' => 1,
-                'fromSheet' => 2,
-                'toSheet' => 3,
-                'slot' => 9,
-                'spot' => 8,
-            ],
-        ]);
+        $lastVersion->getVersion()->willReturn(
+            [
+                1 => [
+                    'request' => 1,
+                    'fromSheet' => 2,
+                    'toSheet' => 3,
+                    'slot' => 9,
+                    'spot' => 8,
+                ],
+            ]
+        )
+        ;
         $lastVersion->getEvent()->willReturn($event->reveal());
         $lastVersion->getUser()->willReturn($user->reveal());
         $event->getTimeZone()->willReturn('Europe/Paris');
@@ -191,11 +211,13 @@ class DiffVerbalizerTest extends TestCase
         $this->sheetRepository
             ->getSheetsByUserAndEvent($user->reveal(), $event->reveal())
             ->shouldBeCalled()
-            ->willReturn([2 => $userSheet->reveal()]);
+            ->willReturn([2 => $userSheet->reveal()])
+        ;
         $this->sheetRepository
             ->findByIds([4 => 4])
             ->shouldBeCalled()
-            ->willReturn([4 => $sheetMet1->reveal()]);
+            ->willReturn([4 => $sheetMet1->reveal()])
+        ;
         $this->translator
             ->trans(
                 DiffVerbalizer::TRANSLATION_MEETING_ADDED,
@@ -209,7 +231,8 @@ class DiffVerbalizerTest extends TestCase
                 $locale
             )
             ->shouldBeCalled()
-            ->willReturn('Meeting added with sheet title given 1 the 10/10/2017 at 12:00 in Spot Ref');
+            ->willReturn('Meeting added with sheet title given 1 the 10/10/2017 at 12:00 in Spot Ref')
+        ;
 
         $this->meetingSlotRepository->findByIds([12 => 12])->shouldBeCalled()->willReturn([12 => $slot->reveal()]);
         $this->spotRepository->getSpotsByIds([14 => 14])->shouldBeCalled()->willReturn([14 => $spot->reveal()]);
@@ -227,7 +250,8 @@ class DiffVerbalizerTest extends TestCase
             $this->translator->reveal(),
             $this->sheetRepository->reveal(),
             $this->meetingSlotRepository->reveal(),
-            $this->spotRepository->reveal()
+            $this->spotRepository->reveal(),
+            $this->messageRepository->reveal()
         );
         $result = $diffVerbalizer->verbalizeDiff($lastVersion->reveal(), $currentVersion, $locale);
 
@@ -266,22 +290,25 @@ class DiffVerbalizerTest extends TestCase
                 'spot' => 14,
             ],
         ];
-        $lastVersion->getVersion()->willReturn([
-            1 => [
-                'request' => 1,
-                'fromSheet' => 2,
-                'toSheet' => 3,
-                'slot' => 10,
-                'spot' => 8,
-            ],
-            2 => [
-                'request' => 2,
-                'fromSheet' => 2,
-                'toSheet' => 4,
-                'slot' => 12,
-                'spot' => 16,
-            ],
-        ]);
+        $lastVersion->getVersion()->willReturn(
+            [
+                1 => [
+                    'request' => 1,
+                    'fromSheet' => 2,
+                    'toSheet' => 3,
+                    'slot' => 10,
+                    'spot' => 8,
+                ],
+                2 => [
+                    'request' => 2,
+                    'fromSheet' => 2,
+                    'toSheet' => 4,
+                    'slot' => 12,
+                    'spot' => 16,
+                ],
+            ]
+        )
+        ;
         $lastVersion->getEvent()->willReturn($event->reveal());
         $lastVersion->getUser()->willReturn($user->reveal());
         $event->getTimeZone()->willReturn('Europe/Paris');
@@ -295,11 +322,13 @@ class DiffVerbalizerTest extends TestCase
         $this->sheetRepository
             ->getSheetsByUserAndEvent($user->reveal(), $event->reveal())
             ->shouldBeCalled()
-            ->willReturn([2 => $userSheet->reveal()]);
+            ->willReturn([2 => $userSheet->reveal()])
+        ;
         $this->sheetRepository
             ->findByIds([3 => 3, 4 => 4])
             ->shouldBeCalled()
-            ->willReturn([3 => $sheetMet1->reveal(), 4 => $sheetMet2->reveal()]);
+            ->willReturn([3 => $sheetMet1->reveal(), 4 => $sheetMet2->reveal()])
+        ;
         $this->translator
             ->trans(
                 DiffVerbalizer::TRANSLATION_MEETING_MOVED,
@@ -313,7 +342,8 @@ class DiffVerbalizerTest extends TestCase
                 $locale
             )
             ->shouldBeCalled()
-            ->willReturn('Meeting moved with sheet title given 1 the 10/10/2017 at 12:00 in Spot Ref 1');
+            ->willReturn('Meeting moved with sheet title given 1 the 10/10/2017 at 12:00 in Spot Ref 1')
+        ;
         $this->translator
             ->trans(
                 DiffVerbalizer::TRANSLATION_MEETING_MOVED,
@@ -327,7 +357,8 @@ class DiffVerbalizerTest extends TestCase
                 $locale
             )
             ->shouldBeCalled()
-            ->willReturn('Meeting moved with sheet title given 2 the 10/10/2017 at 16:30 in Spot Ref 2');
+            ->willReturn('Meeting moved with sheet title given 2 the 10/10/2017 at 16:30 in Spot Ref 2')
+        ;
 
         $this->meetingSlotRepository
             ->findByIds([9 => 9, 12 => 12])
@@ -363,7 +394,8 @@ class DiffVerbalizerTest extends TestCase
             $this->translator->reveal(),
             $this->sheetRepository->reveal(),
             $this->meetingSlotRepository->reveal(),
-            $this->spotRepository->reveal()
+            $this->spotRepository->reveal(),
+            $this->messageRepository->reveal()
         );
         $result = $diffVerbalizer->verbalizeDiff($lastVersion->reveal(), $currentVersion, $locale);
 
@@ -403,22 +435,25 @@ class DiffVerbalizerTest extends TestCase
                 'spot' => 14,
             ],
         ];
-        $lastVersion->getVersion()->willReturn([
-            1 => [
-                'request' => 1,
-                'fromSheet' => 2,
-                'toSheet' => 3,
-                'slot' => 10,
-                'spot' => 8,
-            ],
-            2 => [
-                'request' => 2,
-                'fromSheet' => 2,
-                'toSheet' => 5,
-                'slot' => 12,
-                'spot' => 16,
-            ],
-        ]);
+        $lastVersion->getVersion()->willReturn(
+            [
+                1 => [
+                    'request' => 1,
+                    'fromSheet' => 2,
+                    'toSheet' => 3,
+                    'slot' => 10,
+                    'spot' => 8,
+                ],
+                2 => [
+                    'request' => 2,
+                    'fromSheet' => 2,
+                    'toSheet' => 5,
+                    'slot' => 12,
+                    'spot' => 16,
+                ],
+            ]
+        )
+        ;
         $lastVersion->getEvent()->willReturn($event->reveal());
         $lastVersion->getUser()->willReturn($user->reveal());
         $event->getTimeZone()->willReturn('Europe/Paris');
@@ -433,11 +468,13 @@ class DiffVerbalizerTest extends TestCase
         $this->sheetRepository
             ->getSheetsByUserAndEvent($user->reveal(), $event->reveal())
             ->shouldBeCalled()
-            ->willReturn([2 => $userSheet->reveal()]);
+            ->willReturn([2 => $userSheet->reveal()])
+        ;
         $this->sheetRepository
             ->findByIds([3 => 3, 4 => 4, 5 => 5])
             ->shouldBeCalled()
-            ->willReturn([3 => $sheetMet1->reveal(), 4 => $sheetMet2->reveal(), 5 => $oldSheetMet->reveal()]);
+            ->willReturn([3 => $sheetMet1->reveal(), 4 => $sheetMet2->reveal(), 5 => $oldSheetMet->reveal()])
+        ;
         $this->translator
             ->trans(
                 DiffVerbalizer::TRANSLATION_MEETING_MOVED,
@@ -451,7 +488,8 @@ class DiffVerbalizerTest extends TestCase
                 $locale
             )
             ->shouldBeCalled()
-            ->willReturn('Meeting moved with sheet title given 1 the 10/10/2017 at 12:00 in Spot Ref 1');
+            ->willReturn('Meeting moved with sheet title given 1 the 10/10/2017 at 12:00 in Spot Ref 1')
+        ;
         $this->translator
             ->trans(
                 DiffVerbalizer::TRANSLATION_MEETING_ADDED,
@@ -465,7 +503,8 @@ class DiffVerbalizerTest extends TestCase
                 $locale
             )
             ->shouldBeCalled()
-            ->willReturn('Meeting added with sheet title given 2 the 10/10/2017 at 16:30 in Spot Ref 2');
+            ->willReturn('Meeting added with sheet title given 2 the 10/10/2017 at 16:30 in Spot Ref 2')
+        ;
         $this->translator
             ->trans(
                 DiffVerbalizer::TRANSLATION_MEETING_DELETED,
@@ -474,7 +513,28 @@ class DiffVerbalizerTest extends TestCase
                 $locale
             )
             ->shouldBeCalled()
-            ->willReturn('Meeting removed with old sheet title');
+            ->willReturn('Meeting removed with old sheet title')
+        ;
+        $this->translator
+            ->trans(
+                DiffVerbalizer::TRANSLATION_MEETING_CHANGED_WITH_MESSAGE,
+                ['%message%' => 'I had to delete'],
+                DiffVerbalizer::TRANSLATION_DOMAIN,
+                $locale
+            )
+            ->shouldBeCalled()
+            ->willReturn('Message: I had to delete')
+        ;
+        $this->translator
+            ->trans(
+                DiffVerbalizer::TRANSLATION_MEETING_CHANGED_WITH_MESSAGE,
+                ['%message%' => 'I like to move it move it'],
+                DiffVerbalizer::TRANSLATION_DOMAIN,
+                $locale
+            )
+            ->shouldBeCalled()
+            ->willReturn('Message: I like to move it move it')
+        ;
 
         $this->meetingSlotRepository
             ->findByIds([9 => 9, 12 => 12])
@@ -496,17 +556,43 @@ class DiffVerbalizerTest extends TestCase
             ->willReturn(true)
         ;
 
+        $request1 = $this->prophesize(Request::class);
+        $request1->getId()->willReturn(1);
+
+        $request2 = $this->prophesize(Request::class);
+        $request2->getId()->willReturn(2);
+
+        $request1message = $this->prophesize(Message::class);
+        $request1message->getContent()
+            ->willReturn('I like to move it move it')
+        ;
+        $request1message->getRequest()->willReturn($request1->reveal());
+        $request1message->getFrom()->willReturn($sheetMet2->reveal());
+
+        $request2message = $this->prophesize(Message::class);
+        $request2message->getContent()
+            ->willReturn('I had to delete')
+        ;
+        $request2message->getRequest()->willReturn($request2->reveal());
+        $request2message->getFrom()->willReturn($sheetMet2->reveal());
+
+        $this->messageRepository->getUpdateOrDeleteReasonMessageFromRequestIds([2, 1])
+            ->shouldBeCalled()
+            ->willReturn([$request1message->reveal(), $request2message->reveal()])
+        ;
+
         $diffVerbalizer = new DiffVerbalizer(
             $this->diffChecker->reveal(),
             $this->translator->reveal(),
             $this->sheetRepository->reveal(),
             $this->meetingSlotRepository->reveal(),
-            $this->spotRepository->reveal()
+            $this->spotRepository->reveal(),
+            $this->messageRepository->reveal()
         );
         $result = $diffVerbalizer->verbalizeDiff($lastVersion->reveal(), $currentVersion, $locale);
 
         $this->assertEquals(
-            "Meeting removed with old sheet title\nMeeting added with sheet title given 2 the 10/10/2017 at 16:30 in Spot Ref 2\nMeeting moved with sheet title given 1 the 10/10/2017 at 12:00 in Spot Ref 1",
+            "Meeting removed with old sheet title Message: I had to delete\nMeeting added with sheet title given 2 the 10/10/2017 at 16:30 in Spot Ref 2\nMeeting moved with sheet title given 1 the 10/10/2017 at 12:00 in Spot Ref 1 Message: I like to move it move it",
             $result
         );
     }